set_operator -> write_operator, Operator fixes, first steps to init
Renamed set_operator to write_operator. I wanted to overload __setattr__ so that whenever something in file_operator was changed, it would automatically be written to the pickle file. But, that would involve creating a wrapper around the Operator, whcih I won't do yet.
I've also added a method, destroy_repo(), which deletes everything in the /.git/ folder except for the /.git/gud/ folder. Running git init on the folder at this point still preserves the /gud/ folder.
Loading level 1-2 and 1-1 after loading 0-3 (init level) work fine. But for some reason, a strange error is occuring at when loading 2-2 after 0-3.